A simplest train file for mnist added.

+++ paddle/api/PaddleAPI.h
@@ -799,7 +799,7 @@ public:
 
   void finishPass();
 
-  PassType startBatch(int64_t batchSize);
+  PassType startBatch(size_t batchSize);
 
   void finishBatch(float cost);
